cabinet-quoting-system/database/scripts/requirements.txt
# Python requirements for Cabinet Quoting System database scripts
psycopg2-binary>=2.9.5